% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAB_plot.R
\name{JAB_plot}
\alias{JAB_plot}
\title{Plots JAB as a function of the p-value}
\usage{
JAB_plot(n, BF = 1, method = "JAB", upper = 1)
}
\arguments{
\item{n}{Sample size. A positive numeric vector.}

\item{BF}{Bayes factor you would like to match. 1 to avoid Lindley's Paradox, 3 to achieve moderate evidence and 10 to achieve strong evidence.}

\item{method}{Used for the choice of 'b'. Currently one of:
\itemize{
\item "JAB": this choice of b produces Jeffreys' approximate BF (Wagenmakers, 2022)
\item "min": uses the minimal training sample for the prior (Gu et al., 2018)
\item "robust": a robust version of "min" that prevents too small b (O'Hagan, 1995)
\item "balanced": this choice of b balances the type I and type II errors (Gu et al., 2016)
}}

\item{upper}{The upper limit for the range of realistic effect sizes. Only relevant when method="balanced". Defaults to 1 such that the range of realistic effect sizes is uniformly distributed between 0 and 1, U(0,1).}
}
\value{
Prints a plot.
}
\description{
Plots JAB as a function of the p-value
}
\examples{
# Plot JAB as function of the p-value for a sample size of 2000
JAB_plot(2000)
}
