% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JABp.R
\name{JABp}
\alias{JABp}
\title{Title}
\usage{
JABp(n, p, z = TRUE, df = NULL, method = "JAB")
}
\arguments{
\item{n}{Sample size.}

\item{p}{The p-value.}

\item{z}{Is the p-value based on a z- or t-statistic? TRUE if z.}

\item{df}{If z=FALSE, provide the degrees of freedom for the t-statistic.}

\item{method}{Used for the choice of 'b'. Currently one of:
\itemize{
\item "JAB": this choice of b produces Jeffery's approximate BF
\item "min": uses the minimal training sample for the prior (Gu et al., '17)
\item "robust": a robust version of "min" that prevents too small b (O'Hagan, '95)
\item "balanced": this choice of b balances the type I and type II errors
}}
}
\value{
A numeric value for the BF in favour of H1.
}
\description{
Title
}
\examples{
#' # Transform a p-value of 0.007038863 from a z-test into JAB
# using a sample size of 200.
JABt(200, 2.695)
}
