% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JABt.R
\name{JABt}
\alias{JABt}
\title{Transforms a t-statistic into Jeffreys' approximate Bayes factor}
\usage{
JABt(n, t, method = "JAB", p = NULL)
}
\arguments{
\item{n}{Sample size}

\item{t}{The t-statistic}

\item{method}{Used for the choice of 'b', currently one of "JAB" or "JAB adj"}

\item{p}{Number of parameters in the model, only required if method="JAB adj"}
}
\value{
A numeric value for the BF in favour of H1
}
\description{
Transforms a t-statistic into Jeffreys' approximate Bayes factor
}
\examples{
# Transform a t-statistic of 2.695 computed based on a sample size of 200 into JAB
JABt(200, 2.695)
}
