% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_alpha.R
\name{set_alpha}
\alias{set_alpha}
\title{Set the alpha level based on sample size for coefficients in a  regression models.}
\usage{
set_alpha(n, p = 1, evidence = 1, tstat = TRUE, plotprint = FALSE)
}
\arguments{
\item{n}{The sample size.}

\item{p}{The number of parameters to be tested.}

\item{evidence}{Desired level of evidence: 1 to avoid the Lindley Paradox, 3 to achieve moderate evidence and 10 to achieve strong evidence.}

\item{tstat}{Should the t-statistic or the z-statistic be used? Set as TRUE in the case of linear regression.}

\item{plotprint}{If true prints a plot relating Bayes factors and p-values.}
}
\value{
numeric alpha level required achieve the desired level of evidence.
}
\description{
Set the alpha level based on sample size for coefficients in a  regression models.
}
\section{References}{

Taylor, L. & Wulff, J.N. (2022). Let alpha depend on n: A Bayesian-frequentist compromise by using lower alpha levels in larger sample size
}

\examples{
## Set alpha such that the Lindley paradox is avoided for the test of
## a single coefficient in a linear regression model with 10 regressors
## with a sample size of 1000.

set_alpha(1000, p = 10, evidence = 1)

## Set alpha such to achieve moderate evidence if the null is rejected
## for the test of a single coefficient a logistic regression model with
## 5 regressors and a sample size of 1000. Also print a plot relating
## the Bayes factor to the p-value.

set_alpha(1000, p = 5, evidence=3, tstat = FALSE, plotprint = TRUE)
}
