% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaN.R
\name{alphaN}
\alias{alphaN}
\title{Title}
\usage{
alphaN(n, BF = 1, method = "JAB", p = NULL)
}
\arguments{
\item{n}{Sample size}

\item{BF}{Bayes factor you would like to match. 1 to avoid the Lindley Paradox, 3 to achieve moderate evidence and 10 to achieve strong evidence.}

\item{method}{used for the choice of 'b', currently one of "JAB" or "JAB adj"}

\item{p}{number of parameters in the model, only required if method="JAB adj"}
}
\value{
numeric alpha level required to achieve the desired level of evidence.
}
\description{
Title
}
\section{References}{

Taylor, L. & Wulff, J.N. (2022). Let alpha depend on n: A Bayesian-frequentist compromise by using lower alpha levels in larger sample size
}

\examples{

# Plot of alpha level as a function of n
seqN <- seq(50, 1000, 1)
plot(seqN, alphaN(seqN), type = "l")


}
