% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaN.R
\name{alphaN}
\alias{alphaN}
\title{Set the alpha level based on sample size for coefficients in a regression models.}
\usage{
alphaN(n, BF = 1, method = "JAB")
}
\arguments{
\item{n}{Sample size}

\item{BF}{Bayes factor you would like to match. 1 to avoid Lindley's Paradox, 3 to achieve moderate evidence and 10 to achieve strong evidence.}

\item{method}{Used for the choice of 'b'. Currently one of:
\itemize{
\item "JAB": this choice of b produces Jeffery's approximate BF
\item "min": uses the minimal training sample for the prior (Gu et al., '17)
\item "robust": a robust version of "min" that prevents too small b (O'Hagan, '95)
\item "balanced": this choice of b balances the type I and type II errors
}}
}
\value{
Numeric alpha level required to achieve the desired level of evidence.
}
\description{
Set the alpha level based on sample size for coefficients in a regression models.
}
\section{References}{

Gu et al. (2018). Approximated adjusted fractional Bayes factors: A general method for testing informative hypotheses. The British Journal of Mathematical and Statistical Psychology, 71(2). \cr
\cr
O’Hagan, A. (1995). Fractional Bayes Factors for Model Comparison. Journal of the Royal Statistical Society. Series B (Methodological), 57(1), 99–138. \cr
\cr
Wulff & Taylor (2023). How and why alpha should depend on sample size: A Bayesian-frequentist compromise for significance testing. PsyArXiv. 10.31234/osf.io/3cbh7
}

\examples{
# Plot of alpha level as a function of n
seqN <- seq(50, 1000, 1)
plot(seqN, alphaN(seqN), type = "l")
}
