% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaN_plot.R
\name{alphaN_plot}
\alias{alphaN_plot}
\title{Creates a plot of alpha as function of sample size for each of the four prior options}
\usage{
alphaN_plot(BF = 1, max = 10000)
}
\arguments{
\item{BF}{Bayes factor you would like to match. 1 to avoid Lindley's Paradox, 3 to achieve moderate evidence and 10 to achieve strong evidence.}

\item{max}{The maximum number of sample size. Defaults to 10,000.}
}
\value{
Prints a plot.
}
\description{
Creates a plot of alpha as function of sample size for each of the four prior options
}
\examples{
# Plot of alpha level as a function of n for a Bayes factor of 3
alphaN_plot(BF = 3)
}
