% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAB.R
\name{JAB}
\alias{JAB}
\title{Transforms a t-statistic from a glm or lm object into Jeffreys' approximate Bayes factor}
\usage{
JAB(glm_obj, covariate, method = "JAB")
}
\arguments{
\item{glm_obj}{a glm or lm object.}

\item{covariate}{the name of the covariate that you want a BF for as a string.}

\item{method}{used for the choice of 'b', currently one of "JAB" or "JAB adj".}
}
\value{
A numeric value for the BF in favour of H1.
}
\description{
Transforms a t-statistic from a glm or lm object into Jeffreys' approximate Bayes factor
}
\examples{
# Simulate data

## Sample size
n <- 200

## Regressors
Z1 <- runif(n, -1, 1)
Z2 <- runif(n, -1, 1)
Z3 <- runif(n, -1, 1)
Z4 <- runif(n, -1, 1)
X <- runif(n, -1, 1)

## Error term
U <- rnorm(n, 0, 0.5)

## Outcome
Y <- X/sqrt(n) + U

# Run a GLM
LM <- glm(Y ~ X + Z1 + Z2 + Z3 + Z4)

# Compute JAB for "X" based on the regression results
JAB(LM, "X")

# Compute JAB adj
JAB(LM, "X", method = "JAB adj")
}
