% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAB_plot.R
\name{JAB_plot}
\alias{JAB_plot}
\title{Plots JAB as a function of the p-value}
\usage{
JAB_plot(n, BF = 1, method = "JAB", p = NULL)
}
\arguments{
\item{n}{Sample size}

\item{BF}{Bayes factor you would like to match. 1 to avoid the Lindley Paradox, 3 to achieve moderate evidence and 10 to achieve strong evidence.}

\item{method}{Used for the choice of 'b', currently one of "JAB" or "JAB adj"}

\item{p}{Number of parameters in the model, only required if method="JAB adj"}
}
\value{
Prints a plot
}
\description{
Plots JAB as a function of the p-value
}
\examples{
# Plot JAB as function of the p-value for a sample size of 2000
JAB_plot(2000)
}
