% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JABt.R
\name{JABt}
\alias{JABt}
\title{Transforms a t-statistic into Jeffreys' approximate Bayes factor}
\usage{
JABt(n, t, method = "JAB")
}
\arguments{
\item{n}{Sample size.}

\item{t}{The t-statistic.}

\item{method}{Used for the choice of 'b'. Currently one of:
\itemize{
\item "JAB": this choice of b produces Jeffery's approximate BF
\item "min": uses the minimal training sample for the prior (Gu et al., '17)
\item "robust": a robust version of "min" that prevents too small b (O'Hagan, '95)
\item "balanced: this choice of b balances the type I and type II errors
}}
}
\value{
A numeric value for the BF in favour of H1.
}
\description{
Transforms a t-statistic into Jeffreys' approximate Bayes factor
}
\examples{
# Transform a t-statistic of 2.695 computed based on a sample size of 200 into JAB
JABt(200, 2.695)
}
